076 识别磁盘

+++ src/include/onix/io.h
@@ -6,7 +6,7 @@
 // 输入一个字节
 extern u8 inb(u16 port);
 // 输入一个字
-extern u8 inw(u16 port);
+extern u16 inw(u16 port);
 
 // 输出一个字节
 extern void outb(u16 port, u8 value);
